Move fn GetUsableSpace to common utils
