+++ pkg/util/util.go
@@ -430,3 +430,12 @@ func AppendZeroWithWrite(outFile *os.File, start, length int64) error {
 	}
 	return nil
 }
+
+// GetUsableSpace calculates space to use taking file system overhead into account
+func GetUsableSpace(filesystemOverhead float64, availableSpace int64) int64 {
+	// +1 always rounds up.
+	spaceWithOverhead := int64(math.Ceil((1 - filesystemOverhead) * float64(availableSpace)))
+	// qemu-img will round up, making us use more than the usable space.
+	// This later conflicts with image size validation.
+	return RoundDown(spaceWithOverhead, DefaultAlignBlockSize)
+}
